requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
geoserver-restconfig==2.0.4.8
netCDF4==1.5.4
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.0
boto3==1.17.93
Shapely==1.7.1
python-slugify==5.0.2
psycopg2==2.8.6
click==8.0.1
websocket-client==1.1.0
pandas==1.1.5